Define fluxes needed for the ocean coupling
In particular, these are: the non-solar heat flux (Q_NO_SUN), the
short-wave flux (Q_SW_OCN), the fresh-water flux (FWFLUX), and the salt
flux (SFLUX), all at the ocean surface. I also defined the short wave
flux through the base of the ice, but it's zero for now.

This is an initial commit. It compiles, but is untested and still a bit
messy.