Modified the area function in dynamics/ParametricMesh.cpp:
The element area is computed by transforming the reference element K = [0,1]^2 onto the element T
Hence, Area(T) = \int_T dx = \int_K J(z) dz

The integral is computed with Gauss-Quadrature

Strangely, 1-point Gauss seems to be exact. But I am not sure if this must be the case.
As J = det( nabla A ) I expect that 2 points must be used. A is bi-linear, hence, nabla A
has linear parts and det( nabla A) can be quadratic.
I am using 2 to make it safe :-)

In Spherical Coordinates, the cosine of the lat must be added. This increases the error
Machine precision is only achieved for 3 GP. I propose to use only two, which still gives
10^-9 rel. error.

Added test-case dynamics/tests/ParametricMeshArea_test:

1)
- generates a regular Cartesian mesh of a domain dx x dy
- a) checks that the area of each element is positive
- b) checks that the summed area matches dx x dy up to machine precision
- disturbes all inner mesh nodes by 20% of element size and repeats tests 1a) and 1b)

2)
- generates a spherical mesh of the 'world' [-pi,pi] x [-pi/2, pi/2]
- a) checks that the area of each element is positive
- b) checks that the summed area matches the total surface, 4 Pi R^2
- both tests up to rel. error of 10^-9
- repeat 2a) and 2b) on disturbed elements