Write SST and SSS to restart file
Otherwise, the model can't possibly start from one. The problem was that
TOPAZOcean didn't have a getStatePrognostic (or get stateDiagnostic,
for that matter). Calling getStatePrognostic in SlabOcean, seemed like
the thing to do, but this wanted to call the fields sst_slab and
sss_slab, which is not recognised as a valid restart field by the rest
of the model. So, I renamed those to the canonical sstName and sssName.